reorganized components folder and updated docs
