+++ front-end/src/components/EditorContainer.jsx
@@ -1,21 +1,21 @@
 import { useCallback, useState } from 'react'
 import MediaEntry from './MediaEntry'
-import ImageEditor from './ImageEditor'
+import ImageEditor from './image/ImageEditor'
 import FilterMain from './FilterMain'
-import PresetFilters from './PresetFilters'
-import PresetSizes from './PresetSizes'
-import AddText from './AddText'
-import ColorFilters from './ColorFilters'
-import GifEditor from './GifEditor'
+import PresetFilters from './image/PresetFilters'
+import PresetSizes from './image/PresetSizes'
+import AddText from './image/AddText'
+import ColorFilters from './image/ColorFilters'
+import GifEditor from './gif/GifEditor'
 import useMediaSelection, { MEDIA_SELECTION_CODES } from '../hooks/useMediaSelection'
 import useGifConversion from '../hooks/useGifConversion'
 import useImageEditingSession from '../hooks/useImageEditingSession'
 import { convertBackendImageResultToLocalMedia } from '../services/backendImageService'
 import CameraCapture from './CameraCapture'
 import PhotoPreview from './PhotoPreview'
-import VideoPresetFilters from './VideoPresetFilters'
-import GifFilterMain from './GifFilterMain'
-import GifToolPlaceholder from './GifToolPlaceholder'
+import VideoPresetFilters from './gif/VideoPresetFilters'
+import GifFilterMain from './gif/GifFilterMain'
+import GifToolPlaceholder from './gif/GifToolPlaceholder'
 
 const SCREENS = {
   EDITOR: 'editor',
