+++ front-end/src/components/gif/GifEditor.jsx
