+++ front-end/src/components/gif/GifToolPlaceholder.jsx
