+++ front-end/src/components/image/ImageCropper.jsx
