+++ front-end/src/components/image/PresetSizes.jsx
